refactored seed, working on seed unit test
